Move classes to vfs module
