+++ albam/cache.py
@@ -1,0 +1,2 @@
+
+NODES_CACHE = {}
